refactor

+++ build.sh
@@ -1,7 +1,75 @@
 #!/usr/bin/env bash
-set -eux -o pipefail
+set -eu -o pipefail
 
-vagrant destroy --force
-vagrant up
-vagrant halt
-vagrant package --output chaifeng-bento-ubuntu-16.04-docker-17.12.box
+VAGRANT_CLOUD_USER="${VAGRANT_CLOUD_USER:-chaifeng}"
+VAGRANT_VAGRANTFILE="Vagrantfile-bento"
+
+DOCKER_VERSION="${DOCKER_VERSION:-18.03}"
+BENTO_UBUNTU="${BENTO_UBUNTU:-ubuntu-16.04}"
+BENTO_UBUNTU_VERSION="${BENTO_UBUNTU_VERSION:-201803.24.0}"
+
+export DOCKER_VERSION BENTO_UBUNTU_VERSION BENTO_UBUNTU VAGRANT_VAGRANTFILE
+
+DOCKER="docker-$DOCKER_VERSION"
+BOXNAME="${BENTO_UBUNTU}-${DOCKER}"
+BOXFILE="${VAGRANT_CLOUD_USER}-bento-${BENTO_UBUNTU}-docker-${DOCKER_VERSION}.box"
+
+if [[ ! -f "$BOXFILE" ]]; then
+  echo "Create ${BOXNAME}"
+  vagrant destroy --force
+  vagrant up
+  vagrant halt
+  vagrant package --output "$BOXFILE"
+else
+  echo "Box '${BOXNAME} is already created.'"
+fi
+
+echo ""
+echo "Create a new box."
+curl \
+  --header "Content-Type: application/json" \
+  --header "Authorization: Bearer $VAGRANT_CLOUD_TOKEN" \
+  "https://app.vagrantup.com/api/v1/boxes" \
+  --data "{ \"box\": { \
+              \"username\": \"${VAGRANT_CLOUD_USER}\" \
+              , \"name\": \"${BOXNAME}\" \
+              , \"is_private\": false \
+              , \"short_description\": \
+                   \"Ubuntu ${BENTO_UBUNTU#ubuntu-} with Docker CE ${DOCKER_VERSION}, based on Bento/${BENTO_UBUNTU}\" \
+              } \
+          }"
+
+echo ""
+echo "Create a new version"
+curl \
+    --header "Content-Type: application/json" \
+    --header "Authorization: Bearer $VAGRANT_CLOUD_TOKEN" \
+    "https://app.vagrantup.com/api/v1/box/${VAGRANT_CLOUD_USER}/${BOXNAME}/versions" \
+    --data "{ \"version\": { \"version\": \"201803.24.0\" } }"
+
+echo ""
+echo "Create a new provider"
+curl \
+  --header "Content-Type: application/json" \
+  --header "Authorization: Bearer $VAGRANT_CLOUD_TOKEN" \
+  "https://app.vagrantup.com/api/v1/box/${VAGRANT_CLOUD_USER}/${BOXNAME}/version/$BENTO_UBUNTU_VERSION/providers" \
+  --data "{ \"provider\": { \"name\": \"virtualbox\" } }"
+
+echo ""
+echo "Prepare the provider for upload/get an upload URL"
+response="$(curl \
+  --header "Authorization: Bearer $VAGRANT_CLOUD_TOKEN" \
+  "https://app.vagrantup.com/api/v1/box/${VAGRANT_CLOUD_USER}/${BOXNAME}/version/$BENTO_UBUNTU_VERSION/provider/virtualbox/upload")"
+
+echo "Extract the upload URL from the response"
+upload_path="$(echo "$response" | jq -r .upload_path)"
+
+echo "Perform the upload"
+curl "$upload_path" --request PUT --upload-file "$BOXFILE"
+
+echo "Release the version"
+curl \
+  --header "Authorization: Bearer $VAGRANT_CLOUD_TOKEN" \
+  "https://app.vagrantup.com/api/v1/box/${VAGRANT_CLOUD_USER}/${BOXNAME}/version/$BENTO_UBUNTU_VERSION/release" \
+  --request PUT
+
